Moved rratrap check DM span functions to spio.py
